pyproject.toml
[project]
name = "support-ticket-triage-agent"
version = "0.1.0"
description = "AI agent that triages customer support tickets using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "rank-bm25>=0.2",
    "numpy>=1.26",
    "pandas>=2.0",
]

[project.scripts]
triage = "triage.main:main"

[project.optional-dependencies]
gemini = ["langchain-google-genai>=2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/triage"]
